feat(completion): integrate colorful-menu configuration into blink
Move the colorful-menu plugin configuration to be a dependency and configuration block within the blink plugin.

This simplifies the overall plugin setup by removing the need for a separate colorful-menu plugin file and consolidates the completion menu configuration under the main completion plugin.

The event was also added to ensure the plugin loads when buffers are read or created.

Signed-off-by: Yanhu Cao <gmayyyha@gmail.com>